The PDF of supplementary information
The complete version of the PDF of supplementary information is larger than 25 MB, which is unable to upload to the Github. So a minor version is provided, if the figures in the PDF are not clear enough, please use the folder of "supplementary information figures". The code in it can generate the original figures.